Fix 32/64-bit issues with QFile::map/unmap() on *nix systems
We would previously silently ignore overflows in 32-bit systems and not
properly support 64-bit offsets in systems that support it because of
integer overflow.

There was also a problem that could prevent unmap from succeeding,
because we were passing the wrong length argument.

Task-number: QT-1594

+++ src/corelib/io/qfsfileengine_p.h
@@ -114,7 +114,7 @@ public:
     mutable int cachedFd;
     mutable DWORD fileAttrib;
 #else
-    QHash<uchar *, QPair<int /*offset*/, int /*handle|len*/> > maps;
+    QHash<uchar *, QPair<int /*offset % PageSize*/, size_t /*length + offset % PageSize*/> > maps;
     mutable QT_STATBUF st;
 #endif
     int fd;
